# [esp32_ble] Store GATTC/GATTS param and small data inline to nearly eliminate heap allocations